requirements.txt
pytest==8.0.0
selenium==4.16.0
requests==2.31.0
allure-pytest==2.13.2
webdriver-manager==4.0.1